Refactor raptorQ module
