+++ src/transport/encoding/raptorq/decoder.rs
@@ -4,26 +4,18 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
+use crate::transport::{encoding::Configurable, Decoder};
+use raptorq::{Decoder as ExtDecoder, EncodingPacket};
 use std::{
     collections::HashMap,
-    convert::TryInto,
     time::{Duration, Instant},
-};
-
-use blake2::{Blake2s, Digest};
-use raptorq::{
-    Decoder, Encoder, EncodingPacket, ObjectTransmissionInformation,
 };
 use tracing::warn;
 
-use crate::encoding::{
-    message::Message, payload::BroadcastPayload, Marshallable,
-};
+use crate::encoding::{message::Message, payload::BroadcastPayload};
 
-use super::Configurable;
+use super::ChunkedPayload;
 
-const DEFAULT_REPAIR_PACKETS_PER_BLOCK: u32 = 15;
-const MAX_CHUNK_SIZE: u16 = 1024;
 const CACHE_DEFAULT_TTL_SECS: u64 = 60;
 const CACHE_PRUNED_EVERY_SECS: u64 = 60 * 5;
 
@@ -45,15 +37,8 @@ impl Configurable for RaptorQDecoder {
     }
 }
 
-pub(crate) struct RaptorQEncoder {}
-
-impl Configurable for RaptorQEncoder {
-    fn configure(_: HashMap<String, String>) -> Self {
-        Self {}
-    }
-}
 enum CacheStatus {
-    Receiving(Decoder, Instant),
+    Receiving(ExtDecoder, Instant),
     Processed(Instant),
 }
 
@@ -68,94 +53,7 @@ impl CacheStatus {
     }
 }
 
-struct ChunkedPayload<'a>(&'a BroadcastPayload);
-
-impl BroadcastPayload {
-    fn bytes(&self) -> Vec<u8> {
-        let mut bytes = vec![];
-        self.marshal_binary(&mut bytes).unwrap();
-        bytes
-    }
-    fn generate_uid(&self) -> [u8; 32] {
-        let mut hasher = Blake2s::new();
-        hasher.update(&self.bytes()[1..]);
-        hasher
-            .finalize()
-            .as_slice()
-            .try_into()
-            .expect("Wrong length")
-    }
-}
-impl<'a> ChunkedPayload<'a> {
-    fn uid(&self) -> [u8; 32] {
-        let mut uid: [u8; 32] = Default::default();
-        uid.copy_from_slice(&self.0.gossip_frame[0..32]);
-        uid
-    }
-
-    fn transmission_info(&self) -> ObjectTransmissionInformation {
-        let mut transmission_info: [u8; 12] = Default::default();
-        transmission_info.copy_from_slice(&self.0.gossip_frame[32..44]);
-        ObjectTransmissionInformation::deserialize(&transmission_info)
-    }
-
-    fn encoded_chunk(&self) -> &[u8] {
-        &self.0.gossip_frame[44..]
-    }
-
-    fn safe_uid(&self) -> [u8; 32] {
-        let mut hasher = Blake2s::new();
-        let uid = &self.0.gossip_frame[0..32];
-        let transmission_info = &self.0.gossip_frame[32..44];
-        hasher.update(uid);
-
-        // Why do we need transmission info?
-        //
-        // Transmission info should be sent over a reliable channel, because
-        // it is critical to decode packets.
-        // Since it is sent over UDP alongside the encoded chunked bytes,
-        // corrupted transmission info can be received.
-        // If the corrupted info is part of the first received chunk, no message
-        // can ever be decoded.
-        hasher.update(transmission_info);
-        hasher
-            .finalize()
-            .as_slice()
-            .try_into()
-            .expect("Wrong length")
-    }
-}
-
-impl super::Encoder for RaptorQEncoder {
-    fn encode<'msg>(&self, msg: Message) -> Vec<Message> {
-        if let Message::Broadcast(header, payload) = msg {
-            let uid = payload.generate_uid();
-            let encoder =
-                Encoder::with_defaults(&payload.gossip_frame, MAX_CHUNK_SIZE);
-            encoder
-                .get_encoded_packets(DEFAULT_REPAIR_PACKETS_PER_BLOCK)
-                .iter()
-                .map(|encoded_packet| {
-                    let mut packet_with_uid = uid.to_vec();
-                    let mut transmission_info =
-                        encoder.get_config().serialize().to_vec();
-                    packet_with_uid.append(&mut transmission_info);
-                    packet_with_uid.append(&mut encoded_packet.serialize());
-                    Message::Broadcast(
-                        header,
-                        BroadcastPayload {
-                            height: payload.height,
-                            gossip_frame: packet_with_uid,
-                        },
-                    )
-                })
-                .collect()
-        } else {
-            vec![msg]
-        }
-    }
-}
-impl super::Decoder for RaptorQDecoder {
+impl Decoder for RaptorQDecoder {
     fn decode(&mut self, message: Message) -> Option<Message> {
         if let Message::Broadcast(header, payload) = message {
             let chunked = ChunkedPayload(&payload);
@@ -171,7 +69,7 @@ impl super::Decoder for RaptorQDecoder {
                 // the received transmission information
                 std::collections::hash_map::Entry::Vacant(v) => {
                     v.insert(CacheStatus::Receiving(
-                        Decoder::new(chunked.transmission_info()),
+                        ExtDecoder::new(chunked.transmission_info()),
                         Instant::now(),
                     ))
                 }
